refactor: move gateway source into packages/gateway; remove src/ and legacy tsconfig
Also adds marked-terminal.d.ts ambient declaration to fix TS7016 error
that surfaced when tui.ts moved into the gateway package's strict build.
