+++ packages/gateway/src/cli.ts
