+++ packages/gateway/src/startup-guard.ts
