+++ packages/gateway/src/types/marked-terminal.d.ts
@@ -1,0 +1,4 @@
+declare module 'marked-terminal' {
+  import { MarkedExtension } from 'marked';
+  export function markedTerminal(options?: Record<string, unknown>): MarkedExtension;
+}
